Lab 7 CPRE 381 Simplified MIPS Processor
Object of this lab was to implement a simplified version of a MIPS processor. A lot of the components were given to me already. Most of it was connecting all the given components and creating test bench cases to show functionalities of the processor. Run the tb_SimplifiedMIPSProcessor for a long period of time in model sim. Supply a 100ns period clock cycle and run it for 10000ns to see a series of operation preformed by the test bench. The Matrix.s and HILO register files are some practice codes that work with floating point numbers in MIPS assembly